Added a new top-level file llvm2ice.cpp, with a 'main' of its own.
This program will serve as the playground for LLVM IR -> ICE conversions. The
Makefile was set up to link it correctly with LLVM. subzerotest will keep
compiling without LLVM for now, so that it doesn't break if one doesn't have
LLVM source and build dirs.

Note that this will work with a binary distribution of LLVM as well (so for
example 3.4 can be donwloaded from llvm.org/releases and untarred), as lon as
the LLVM_*_PATH variables are set accordingly to find the headers and the libs.