Adding the basic ParseIRFile skeleton code to llvm2ice - parse and dump a module
Also adding the ir_samples/ directory where (for now) some LLVM IR samples will
live, along with the C files that clang compiles to them. At a later stage we
may want a more elaborate generation scheme where we use the PNaCl toolchain to
freeze and then thaw the module (so we see the stable PNaCl ABI-version of the
IR).