llvm2ice can now convert ir_samples/emptyfunc.ll to ICE.
A very trivial conversion is supported (only voids, only a single BB, a ret,
etc.), but it's a start.

Also, I took parts of the PNaCl ABI verifier into the repo. Still undecided on
how to arrange it in terms of the repositories. Hopefully the PNaCl ABI verifier
can make it to upstream LLVM some day and then we won't need to carry it
around. Otherwise, until Subzero moves into the PNaCl LLVM fork, we'll have
to carry the relevant parts around. The ABI verifier is not necessarily a
"production feature" of Subzero at this point anyway, so this isn't a big
problem (we only use it for testing and sanity checking).