Add instruction numbering.
This will be used to represent operand live ranges for linear-scan
register allocation.  The instruction number is dumped with the
instruction output.  All instructions get a number, including phi
instructions.

BTW, super annoying to have to pass the Cfg pointer around everywhere.
This is essentially a context that should be available thread-locally
from everywhere.