Two changes/fixes to live range calculation.
First, make the live range of a phi destination start with the first
phi instruction of the block, rather than the first non-phi
instruction.  This may lead to more compactly represented live ranges.
Or maybe not.  The main requirement, I believe, is that the phi source
operands are not considered live during the phi instructions.

Second, for 3-address instructions like "a=b-c" where c's live range
ends, we artifically add 1 to the end of c's live range so that it
appears to intersect a's live range, and the register allocator won't
try to coalesce a and c.  Note that we might be able to commute the
operation and then coalesce the two operands.