Initial lit-based test suite.
Using lit because Subzero will eventually migrate into the PNaCl tree or will
be a LLVM subproject.

The lit test suite setup here is hacky; in a LLVM-based tree, we'll be able to
just use the the existing setup, so the important part is to write the tests
themselves in a manner that resembles LLVM tests. This way there won't be
rewriting to do when the project is moved.