Start more serious tuning/experimenting with linear scan.
Turn off everything to do with the old multi-block register
allocation.  Turn off phi lowering for now.

Start tuning the register allocation policy.  For now, limit to
callee-save registers, and limit to variables with multi-block
lifetime or that are destination of phi instructions.  Actually assign
the results of linear-scan register allocation to the variables.

Improve some of the lowering code sequences when source operands are
known to be in physical registers as a result of linear-scan register
allocation.