Add the build-pnacl-ir.py script.
This script uses tools from a NaCl SDK installation to compile C code to
valid PNaCl bitcode in textual format.

Note: this is *not* "final, translateable PNaCl bitcode", because that also
implies the PNaCl .bc binary format which is incompatible with LLVM. The goal
is to produce LLVM IR that constitutes valid PNaCl bitcode semantically.