Add support for lowering of Call instructions.
This includes the addition of a Kill pseudo-instruction in the
top-level ICE, which represents the killing of registers by a lowered
instruction such as Call.  This is represented by creating tiny live
ranges for a variable precolored to that physical register.

There is still a bug to be fixed - on the "fib" test, for some reason
the register allocator is assigning ecx to a variable that is live
across call sites.