Change IceInst::LiveRangesEnded from SmallBitVector to uint32_t.
As a consequence, we can now mark the end of live range of "only" the
first 32 operands in an instruction.  That restriction shouldn't break
anything, since the contract is that if LiveRangesEnded[i] is true,
then operand i's live range definitely ends, but if LiveRangesEnded[i]
is false, then it may or may not end.

For now, LiveRangesEnded is actually only used for dumping and for the
legacy IceRegManager based lowering.  For the latter, we optimize by
reusing the register of an operand whose live range we know is ending,
and going so far as to commute the operator if that's possible and
helps.

Currently, there are only two instruction types that can possibly
exceed 32 operands: the call instruction, and the FakeKill
instruction.  LiveRangesEnded is not really relevant to FakeKill.