Add -timing option to Subzero to break down the timing of its operation a bit.
Using a separate timer from LLVM's infrastructure just to not get tied into
LLVM's flags for this, and also to enable easy per-function reporting. We may
want to combine at some point.

The main goal here is to see how much of Subzero's time goes on parsing IR,
converting it to ICE and emitting (as opposed to the "meat" of translation).