Fix IceCfgNode linearization.
In convertFunction(), add an initial pass over the basic blocks to
pre-resolve them in bitcode order.  Without this, converting forward
branch instructions may add the branch target to the linearization too
early and mess up the topological sort.  If the topological sort is
off, it slows down convergence of liveness analysis.  For example,
this change reduces the number of passes from 9 to 3 in the bigfunc
example.