Set a more consistent live range in rare cases.
This section of code deals with the rare case where a variable's live
range within a basic block consists of two segments with a gap in the
middle.  This can happen in a loop where the variable is set mid-block
but is a source operand for a Phi instruction at the beginning of the
block.

We set the latter segment's upper bound to be one past the last
instruction of the block to be consistent with other code that does
the same, and it also increases the chances of merging the segment
with the segment in the following block.