Improve the tracking of block-local variables.
Removes the IceVariable::IsMultiblockLife field and consolidates the
logic into IceVariable::DefOrUseNode.  This means that we have to know
what IceCfgNode we're currently in whenever a new IceVariable is
created.  DefOrUseNode is initialized to the creating node, and then
reset to NULL if it is used in a different node.  Also, when
dumping/emitting an IceVariable, the IceOstream keeps track of the
current node for validation purposes.

The reason for this is so that we can accurately deal with
IceVariable-to-IceCfgNode sparsity to improve the performance of
liveness analysis.