Do simple coalescing of stack variables.
One block of space on the stack frame is given to "global" variables
(those used across multiple basic blocks), and no coalescing is done
among them.

Another block of space is given to "local" variables (those with
single-block lifetime).  There is no coalescing within a block, but
the same space is reused for every block.  The space allocated
corresponds to the block needing the most space.

On the bigfunc.pnacl.ll test, the frame size for the x8632 target
reduces from 5604 to 3424 bytes, and the frame size for the x8632fast
target reduces from 19940 to 6220 bytes.

This is only lightly tested so far.