Add separate FileCheck commands for "ICE translation error" messages.
It turns out the CHECK-NOT lines at the end of each file only check
for ICE translation error messages that come after the last CHECK
line, whereas we really want to check the entire output regardless of
the normal CHECK lines.  We fix this by running a separate FileCheck
that just checks for the error message.