Add the --prefix option for testing.
Running e.g. "llvm2ice --prefix=Subzero_" causes all external symbols
to have "Subzero_" prepended to their names.  This makes it easier for
a test harness to do A/B testing of Subzero-translated bitcode against
llc-translated bitcode, using the same bitcode file for both
translators.

One problem/annoyance is that Subzero functions can't make calls to
external functions because of the prefix (with the exception of
math/cast helpers with fixed names), but that's probably not a big
deal for the kind of testing this is meant for.