Fix some issues with narrow integer types.
This keeps types more or less consistent in the lowering.  For
example, an i8 IceVariable on the stack would only use the low-order
byte out of the 4-byte stack slot, and the low 8-bit version of the
physical register would be used.  Also, arithmetic instructions would
operate on the 8-bit operands.

It may be better to overhaul the internal type mechanism.  The idea
would be that after lowering, all IceVariable types would correspond
to natural architectural widths.  For example, all integer variables
in x86-32 would be IceType_i32.  This is already done as part of the
PNaCl ABI at function boundaries, where arguments and return values
are no narrower than IceType_i32.  Load and store instructions would
take care of the widening and narrowing from and to memory.  As long
as implicit widening is done with zero-extension, I think arithmetic
operations would just work out.  The one exception is the ashr
(arithmetic right-shift) instruction, where the sign bit isn't
necessarily copied from bit 31.  But for an ashr e.g. of an i8 by N
bits, this could probably be solved by left-shifting 24 bits and then
arithmetic right-shifting 24+N bits.