Simplify llvm2ice type extraction.
When converting an LLVM type to an ICE type, originally we would
convert a pointer type into its equivalent pointed-to type.  This was
so that the representation and lowering of IceInstLoad and
IceInstStore could know the type of the actual memory operand.  This
happened to work out because Load and Store instructions are the only
places in simplified PNaCl bitcode where pointer types actually show
up, rather than being converted to ints.

This pointer-stripping ends up being unnecessary since the pointed-to
type is available through the data operand of the Load or Store
instruction.

However, making this simplification exposed a bug in the constant
pooling code, where symbolic constants with different immediate
offsets were being combined.