Cleanup of global context and output streams.
Provide separate output streams for dump() versus emit().

Introduce a global context holding things that (mostly) don't change,
such as output streams, verbose flags, and the Target.

Remove the Cfg and other stuff from IceOstream.

Make most of the dump()/emit() routines take a Cfg* argument instead
of an IceOstream.

Remove most of the IceOstream operator<<() overloading.  The main
problem is that IceVariables need Cfg state for dumping/emitting and I
just don't see a way to provide that.

There is still work to do here.  IceTargetLowering and
IceTargetLoweringX8632 are holding Cfg-specific state which is not
appropriate.  The global constant pool needs to be moved into
IceGlobalContext.  Accesses to various IceGlobalContext fields needs
to be synchronized.