Fix the i8 Udiv/Urem emission.
Unfortunately, in this case we have to explicitly use the ah register
with the div instruction.  Until now, we've been able to avoid the
complications from having 2 indepdendent registers (al and ah) that
alias other registers (ax and eax).  The al/ax/eax aliasing is easily
handled by giving them all the same register number, which makes
liveness analysis and register allocation work out.  To deal with ah
(and the remaining 8-bit registers) properly, we'll need to make the
register allocator explicitly aware of register aliasing, and possibly
liveness analysis as well.

The hack here is to explicitly give ah a distinct register number, but
not include it in any of the register allocator's available register
sets.  That way, the register allocator doesn't touch ah except where
the target lowering pre-colors a variable to ah.