Fix calculation of stack offsets for arguments.
Background: During code emission, the current esp offset from its
"home" location is maintained.  Push instructions increment the
offset, and call instructions reset it to 0.  This way, stack
variables can be accurately accessed in the middle of a push sequence
for call.

The problem here is that in the function prolog, saves of callee-save
registers use push instructions, and these push instructions were
inappropriately updating the esp offset.

This bug rarely showed up because:

* The Om1 generator rarely uses callee-save registers (64-bit multiply
being one exception).

* For smaller tests, the O2 generator is aggressive about copying
arguments into registers before the esp adjustment becomes incorrect.

* The esp offset becomes correct after the first call instruction.

The fix (for now) is to add a flag to the push instruction that
suppresses the esp adjustment, and this flag is used for the
callee-save push instructions.